projectBackend/requirements.txt
asgiref==3.5.2
cffi==1.15.0
cryptography==37.0.2
Django==4.0.4
django-cors-headers==3.12.0
django-filter==21.1
django-rest-framework==0.1.0
django-rest-framework-mongoengine==3.4.1
django-rest-knox==4.2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
mongoengine==0.24.1
pycparser==2.21
PyJWT==2.4.0
pymongo==4.1.1
pytz==2022.1
sqlparse==0.4.2
tzdata==2022.1
